refactor into small Free/Cofree interpreter.
