+++ typeclass/protocols/functor.py
@@ -1,5 +1,5 @@
 from typing import TypeVar, Protocol, Callable, Generic, Self
-from dataclass import dataclass
+from dataclasses import dataclass
 
 A = TypeVar("A")
 B = TypeVar("B")
@@ -37,53 +37,3 @@ class Functor(Protocol, Generic[A]):
             Box(2).fmap(lambda x: x * 3) -> Box(6)
         """
         ...
-
-@dataclass
-class FreeFmap:
-    func: Callable[[A], B]
-    value: Functor[A]
-
-def fmap(functor: Functor[A], f: Callable[[A], B]) -> Functor[B]:
-    """
-    Map a function over a functor using external syntax.
-
-    This is the standalone equivalent of `functor.fmap(f)` and matches Haskell's `fmap`.
-
-    Args:
-        f: A function to apply.
-        functor: A Functor instance.
-
-    Returns:
-        A new functor with the function applied.
-    """
-    return FreeFmap(f, Thunk(lambda:functor))
-
-def replace(value: A, functor: Functor[B]) -> Functor[A]:
-    """
-    Replace all values in the functor with the given value.
-
-    Equivalent to Haskell's `<$`.
-
-    Args:
-        value: The replacement value.
-        functor: A Functor whose structure will be preserved.
-
-    Returns:
-        A new functor where every element is replaced with `value`.
-    """
-    return functor.fmap(lambda _: value)
-
-def void(functor: Functor[A]) -> Functor[None]:
-    """
-    Replace all values in the functor with `None`.
-
-    This is equivalent to Haskell's `void`, useful when you're only interested in the
-    effects or structure and not the result.
-
-    Args:
-        functor: A Functor instance.
-
-    Returns:
-        A functor of the same shape with all values replaced by None.
-    """
-    return replace(None, functor)
